Move classes & incorrect total entries count
Move classes to designated folders and also fixed issue when sql for
total num of rows in db returned 1 on every occasion,
